# swagger-sample.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Kindergarden API"
  },
  "servers": [
    {
      "url": "https://serene-virgin-islands-65505.herokuapp.com/"
    }
  ],
  "paths": {
    "/kindergarden": {
      "get": {
        "summary": "List all kindergartens",
        "tags": ["resource1"],
        "responses": {
          "200": {
            "description": "An array of kindergartens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": {
                      "centre_code": "KG1002",
                      "centre_name":
                        "PCF SPARKLETOTS PRESCHOOL @ TANJONG PAGAR-TIONG BAHRU Blk 1E (KN)",
                      "levels_offered": "Playgroup",
                      "programme_hours": "na",
                      "second_languages_offered": "Chinese",
                      "vacancy": "contactcentre",
                      "registration_fee": "16.05",
                      "fees_sc": "na",
                      "fee_pr": "na",
                      "last_updated": "25-Oct-17"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new Kindergarten",
        "tags": ["resource1"],
        "responses": {
          "201": {
            "description": "Null response"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/kindergarden/{center_code}": {
      "get": {
        "summary": "Info for a specific kindergarten",
        "tags": ["resource1"],
        "parameters": [
          {
            "name": "center_code",
            "in": "path",
            "required": true,
            "description": "The center code",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 5,
                    "name": "something"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific item",
        "tags": ["resource1"],
        "parameters": [
          {
            "name": "center_code",
            "in": "path",
            "required": true,
            "description": "The id of the resource to update",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Updated fields for resource",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "id": 2,
                  "name": "something"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 5,
                    "name": "updated name"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resource2": {
      "get": {
        "summary": "List all resource1",
        "tags": ["resource2"],
        "responses": {
          "200": {
            "description": "List of resource objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": {
                      "id": 6,
                      "special": "a different kind of resource"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}